Unified search: expose the v2 <SearchResults> surface on the card @context
The card @context now carries `searchResultsComponent` — the v2
`<SearchResults>` rendering surface — alongside the existing
`prerenderedCardSearchComponent`, which is marked @deprecated. Card authors
opt into the v2 surface; first-party card source migrates to it via the
@context search codemod, after which the deprecated affordance is removed.

The card-facing contract (`SearchResultsComponentSignature` plus its yield
and entry view-model) lives in runtime-common so `CardContext` can type the
member without importing host code; `HydrationMode` and `SearchEntryRendering`
move there too and are re-exported from their host locations. The three root
@context providers expose the host `<SearchResults>` through the new member;
the operator-mode providers inherit it via their context spread.

The cards-facing `Store` interface stays instances-only (no `searchEntries`),
already codified at the type level.
