+++ packages/base/card-api.gts
@@ -73,6 +73,7 @@ import {
   type getCardCollection,
   type Store,
   type PrerenderedCardComponentSignature,
+  type SearchResultsComponentSignature,
   type ErrorEntry,
   type Query,
   type QueryWithInterpolations,
@@ -344,7 +345,18 @@ export interface CardContext<T extends CardDef = CardDef> {
       };
     };
   }>;
+  /**
+   * @deprecated Use {@link CardContext.searchResultsComponent} — the v2
+   * `<SearchResults>` surface. Retained (and still provided) during the
+   * migration window; first-party card source moves to the v2 surface via the
+   * `@context` search codemod, after which this is removed.
+   */
   prerenderedCardSearchComponent: typeof GlimmerComponent<PrerenderedCardComponentSignature>;
+  // The v2 search rendering surface: renders the heterogeneous `search-entry`
+  // stream for a `search-entry`-rooted query — prerendered HTML inert (hydrated
+  // lazily) or a live card — so a card author renders results without ever
+  // branching on prerendered-vs-live. Supersedes `prerenderedCardSearchComponent`.
+  searchResultsComponent: typeof GlimmerComponent<SearchResultsComponentSignature>;
   getCard: getCard<T>;
   getCards: getCards;
   getCardCollection: getCardCollection;
